pyproject.toml
[project]
name = "quest-py"
version = "0.1.0"
description = "Quest Platform - Clean Python Rewrite with Temporal and Pydantic AI Gateway"
requires-python = ">=3.11"
dependencies = [
    # Temporal
    "temporalio>=1.7.0",

    # AI Gateway
    "openai>=1.0.0",  # For Pydantic AI Gateway (OpenAI-compatible)
    "httpx>=0.27.0",

    # Database
    "psycopg[binary]>=3.0.0",
    "asyncpg>=0.29.0",

    # Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Integrations
    "mux-python>=3.0.0",
    "zep-cloud>=2.0.0",

    # Research
    "crawl4ai>=0.4.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
